new bot, 3 days before comp

+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/common/hardware/wrappers/WEncoder.java
@@ -10,8 +10,7 @@ public class WEncoder implements HardwareDevice {
         this.encoder = encoder;
     }
 
-
-    public double getPosition() {
+    public int getPosition() {
         return this.encoder.getPosition();
     }
